Move Parser classes to their own files
